+++ src/dashboard/employer-dashboard/employerNavigation.jsx
@@ -11,11 +11,11 @@ import AddEmployee from './AddEmployee';
 import EmployeeList from './AllEmployeeList';
 import EmployeeManagement from './EmployeeManagement';
 import { collection, getDocs } from 'firebase/firestore';
-import { db } from '../userAuth/firebase';
+import { db } from '@userAuth/firebase';
 import { AppProvider, DashboardLayout } from '@toolpad/core';
-import theme from '../theme/theme';
+import theme from '@theme/theme';
 import { ThemeProvider } from '@emotion/react';
-import logo from "../assets/logo.png";
+import logo from "@assets/logo.png";
 
 
 //This is only for navigation side bar title and icons, it does not have any functionality 
@@ -55,11 +55,21 @@ const NAVIGATION = [
     icon: <PeopleAltIcon />,
     children: [
       {
+        segment: 'addEmployee',
+        title: 'Add Employee',
+        icon: <PersonAddIcon />,
+      },
+      {
         segment: 'employeeList',
         title: 'Employee List',
         icon: <DescriptionIcon />,
       },
     ],
+  },
+  {
+    segment: 'contactList',
+    title: 'Contact List',
+    icon: <ContactsIcon />,
   },
 ];
 
@@ -77,7 +87,7 @@ function useDemoRouter(initialPath) {
   return router;
 }
 
-function EmployeeNavigation() {
+function EmployerNavigation() {
   const [employees, setEmployees] = useState([]);
   const [isLoading, setIsLoading] = useState(true);
   const [error, setError] = useState(null);
@@ -125,11 +135,17 @@ function EmployeeNavigation() {
         // return <EmployeeManagement employees={employees} setEmployees={setEmployees} defaultView="scheduler" />;
       case 'employeeManagement':
         switch(subSegment) {
+          case 'addEmployee':
+            return <EmployeeManagement employees={employees} setEmployees={setEmployees} defaultView="addEmployee" />;
           case 'employeeList':
             return <EmployeeManagement employees={employees} setEmployees={setEmployees} defaultView="list" />;
           default:
             return <EmployeeManagement employees={employees} setEmployees={setEmployees} defaultView="list" />;
         }
+      case 'contactList':
+        return <div>Contact List</div>;
+      default:
+        return <EmployeeManagement employees={employees} setEmployees={setEmployees} defaultView="scheduler" />;
     }
   };
 
@@ -152,4 +168,4 @@ function EmployeeNavigation() {
   );
 }
 
-export default EmployeeNavigation;
+export default EmployerNavigation;
